backend/requirements.txt
googlemaps~=4.10.0
flask~=3.0.1
python-dotenv~=1.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

requests~=2.32.3
urllib3~=2.2.3
beautifulsoup4~=4.12.3